+++ src/ipinfoWrapper.ts
@@ -16,24 +16,6 @@ import {
 } from "./common";
 import VERSION from "./version";
 
-export {
-    Asn,
-    Company,
-    Carrier,
-    Privacy,
-    Abuse,
-    Domains,
-    IPinfo,
-    Prefix,
-    Prefixes6,
-    AsnResponse,
-    MapResponse,
-    BatchResponse,
-    Cache,
-    LruCache
-};
-export { Options } from "lru-cache";
-
 const clientUserAgent = `IPinfoClient/nodejs/${VERSION}`;
 
 export default class IPinfoWrapper {