+++ test-app/index.js
@@ -1,7 +1,13 @@
-let IPinfo = require("node-ipinfo");
+let { IPinfoWrapper, LruCache } = require("node-ipinfo");
+
+let cacheOptions = {
+    max: 5000,
+    maxAge: 24 * 1000 * 60 * 60,
+};
+let cache = new LruCache(cacheOptions);
 
 let token = process.env.IPINFO_TOKEN
-let ipinfo = new IPinfo(token);
+let ipinfo = new IPinfoWrapper(token, cache);
 
 ipinfo.lookupIp("1.1.1.1").then((response) => {
     console.log(response);